Refactor state mutation handling by removing handleStateMutation function
- Deleted the handleStateMutation function from handle-state-mutation.ts.
- Updated insert-transaction-state.test.ts to include tests previously in handle-state-mutation.test.ts, ensuring proper error handling and state management.
- Modified schema.ts to directly use insertTransactionState for handling state deletions, removing the dependency on handleStateMutation.
